% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{analysisResults<-}
\alias{analysisResults<-}
\title{Set analysis results}
\usage{
analysisResults(x) <- value
}
\arguments{
\item{x}{SpatialMetabolic object}

\item{value}{List of results}
}
\description{
Set analysis results
}
