% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{analysisResults}
\alias{analysisResults}
\alias{analysisResults,SpatialMetabolic-method}
\alias{analysisResults<-,SpatialMetabolic-method}
\title{Access analysis results}
\usage{
analysisResults(x)

\S4method{analysisResults}{SpatialMetabolic}(x)

\S4method{analysisResults}{SpatialMetabolic}(x) <- value
}
\arguments{
\item{x}{A SpatialMetabolic object}

\item{value}{List of analysis results}
}
\value{
List of analysis results

Modified SpatialMetabolic object
}
\description{
Access analysis results
}
\section{Methods (by class)}{
\itemize{
\item \code{analysisResults(SpatialMetabolic)}: Access analysis results

\item \code{analysisResults(SpatialMetabolic) <- value}: Set analysis results

}}
