% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_extras.R
\name{animateSpatialMetabolics}
\alias{animateSpatialMetabolics}
\title{Create animated spatial plot}
\usage{
animateSpatialMetabolics(
  object,
  pathways = NULL,
  sample = NULL,
  fps = 10,
  duration = 10
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{pathways}{Pathways to animate through}

\item{sample}{Sample to plot}

\item{fps}{Frames per second}

\item{duration}{Duration in seconds}
}
\value{
gganimate object
}
\description{
Create animated spatial plot
}
