% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{calculateMetabolicScores}
\alias{calculateMetabolicScores}
\title{Calculate metabolic scores}
\usage{
calculateMetabolicScores(object, ...)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{...}{Additional parameters}
}
\description{
Calculate metabolic scores
}
