% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateQCMetrics}
\alias{calculateQCMetrics}
\title{Calculate quality control metrics}
\usage{
calculateQCMetrics(
  object,
  mitochondrial_pattern = "^mt-|^Mt-|^MT-",
  ribosomal_pattern = "^Rp[sl]|^RP[SL]"
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{mitochondrial_pattern}{Pattern to identify mitochondrial genes}

\item{ribosomal_pattern}{Pattern to identify ribosomal genes}
}
\value{
Object with QC metrics added to colData
}
\description{
Calculate quality control metrics
}
