% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareStates.R
\name{compareMetabolicStates,SpatialMetabolic-method}
\alias{compareMetabolicStates,SpatialMetabolic-method}
\title{Compare metabolic states between conditions}
\usage{
\S4method{compareMetabolicStates}{SpatialMetabolic}(
  object,
  condition = "condition",
  ref_group = NULL,
  test_group = NULL,
  test_type = "wilcox",
  features = NULL,
  min_spots = 3,
  spatial_correction = FALSE,
  n_permutations = 1000,
  verbose = TRUE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{condition}{Column name with condition information}

\item{ref_group}{Reference group name}

\item{test_group}{Test group name}

\item{test_type}{Type of test ("wilcox", "t.test", "spatial_lm", "permutation")}

\item{features}{Features to test (default: metabolic scores)}

\item{min_spots}{Minimum spots per group for testing}

\item{spatial_correction}{Whether to correct for spatial autocorrelation}

\item{n_permutations}{Number of permutations for permutation test}

\item{verbose}{Print progress messages}

\item{BPPARAM}{BiocParallel parameters}
}
\value{
Data frame with differential expression results
}
\description{
Compare metabolic states between conditions
}
\examples{
\dontrun{
# Compare metabolic scores between conditions
results <- compareMetabolicStates(spm,
                                 condition = "condition",
                                 ref_group = "WT",
                                 test_group = "SVD")

# Compare specific genes
gene_results <- compareMetabolicStates(spm,
                                      condition = "condition",
                                      ref_group = "WT",
                                      test_group = "SVD",
                                      features = c("Ndufa1", "Atp5a1"),
                                      test_type = "t.test")
}
}
