% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{compareMetabolicStates}
\alias{compareMetabolicStates}
\title{Compare metabolic states}
\usage{
compareMetabolicStates(object, ...)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{...}{Additional parameters}
}
\description{
Compare metabolic states
}
