% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{comparisonGroups<-}
\alias{comparisonGroups<-}
\title{Set comparison groups}
\usage{
comparisonGroups(x) <- value
}
\arguments{
\item{x}{SpatialMetabolic object}

\item{value}{Factor of groups}
}
\description{
Set comparison groups
}
