% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{comparisonGroups}
\alias{comparisonGroups}
\alias{comparisonGroups,SpatialMetabolic-method}
\alias{comparisonGroups<-,SpatialMetabolic-method}
\title{Access comparison groups}
\usage{
comparisonGroups(x)

\S4method{comparisonGroups}{SpatialMetabolic}(x)

\S4method{comparisonGroups}{SpatialMetabolic}(x) <- value
}
\arguments{
\item{x}{A SpatialMetabolic object}

\item{value}{Factor of comparison groups}
}
\value{
Factor of comparison groups

Modified SpatialMetabolic object
}
\description{
Access comparison groups
}
\section{Methods (by class)}{
\itemize{
\item \code{comparisonGroups(SpatialMetabolic)}: Access comparison groups

\item \code{comparisonGroups(SpatialMetabolic) <- value}: Set comparison groups

}}
