% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_extras.R
\name{createMetabolicReport}
\alias{createMetabolicReport}
\title{Create summary report plot}
\usage{
createMetabolicReport(object, save_to = NULL)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{save_to}{File path to save report}
}
\value{
Combined plot object
}
\description{
Create summary report plot
}
