% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpatialMetabolic.R
\name{createSpatialMetabolic}
\alias{createSpatialMetabolic}
\title{Create SpatialMetabolic object from Seurat objects}
\usage{
createSpatialMetabolic(
  seurat_list,
  condition_col = "condition",
  sample_col = "sample",
  bin_size = 8,
  min_features = 200,
  min_spots = 10
)
}
\arguments{
\item{seurat_list}{List of Seurat objects with spatial data}

\item{condition_col}{Column name containing condition information}

\item{sample_col}{Column name containing sample information}

\item{bin_size}{Bin size for Visium HD data (default: 8)}

\item{min_features}{Minimum number of features per spot}

\item{min_spots}{Minimum number of spots per feature}
}
\value{
SpatialMetabolic object
}
\description{
Create SpatialMetabolic object from Seurat objects
}
\examples{
\dontrun{
# Load Seurat objects
wt1 <- Load10X_Spatial("data/WT1/")
wt2 <- Load10X_Spatial("data/WT2/")
svd1 <- Load10X_Spatial("data/SVD1/")
svd2 <- Load10X_Spatial("data/SVD2/")

# Add metadata
wt1$condition <- "WT"
wt1$sample <- "WT1"
wt2$condition <- "WT"
wt2$sample <- "WT2"
svd1$condition <- "SVD"
svd1$sample <- "SVD1"
svd2$condition <- "SVD"
svd2$sample <- "SVD2"

# Create SpatialMetabolic object
seurat_list <- list(wt1, wt2, svd1, svd2)
spm <- createSpatialMetabolic(seurat_list,
                              condition_col = "condition",
                              sample_col = "sample")
}
}
