% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{detectMetabolicGradients}
\alias{detectMetabolicGradients}
\title{Detect metabolic gradients}
\usage{
detectMetabolicGradients(object, ...)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{...}{Additional parameters}
}
\description{
Detect metabolic gradients
}
