% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGradients.R
\name{.buildSpatialGraph}
\alias{.buildSpatialGraph}
\title{Build spatial neighborhood graph}
\usage{
.buildSpatialGraph(object, n_neighbors, graph_type, radius)
}
\description{
Build spatial neighborhood graph
}
\keyword{internal}
