% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetabolicScores.R
\name{.calculateAUCellScores}
\alias{.calculateAUCellScores}
\title{Placeholder for AUCell scoring}
\usage{
.calculateAUCellScores(object, pathway_list, BPPARAM)
}
\description{
Placeholder for AUCell scoring
}
\keyword{internal}
