% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGradients.R
\name{.calculateGearysC}
\alias{.calculateGearysC}
\title{Calculate Geary's C}
\usage{
.calculateGearysC(test_mat, adj_mat, permutations, BPPARAM)
}
\description{
Calculate Geary's C
}
\keyword{internal}
