% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetabolicScores.R
\name{.calculateMeanScores}
\alias{.calculateMeanScores}
\title{Simple mean expression scoring}
\usage{
.calculateMeanScores(object, pathway_list, BPPARAM)
}
\description{
Simple mean expression scoring
}
\keyword{internal}
