% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGradients.R
\name{.calculateMoransI}
\alias{.calculateMoransI}
\title{Calculate Moran's I with permutation testing}
\usage{
.calculateMoransI(test_mat, adj_mat, permutations, BPPARAM)
}
\description{
Calculate Moran's I with permutation testing
}
\keyword{internal}
