% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetabolicScores.R
\name{.calculateSeuratScores}
\alias{.calculateSeuratScores}
\title{Seurat-style scoring implementation}
\usage{
.calculateSeuratScores(object, pathway_list, nbin, ctrl, BPPARAM)
}
\description{
Seurat-style scoring implementation
}
\keyword{internal}
