% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetabolicScores.R
\name{.calculateSSGSEAScores}
\alias{.calculateSSGSEAScores}
\title{Placeholder for ssGSEA scoring}
\usage{
.calculateSSGSEAScores(object, pathway_list, BPPARAM)
}
\description{
Placeholder for ssGSEA scoring
}
\keyword{internal}
