% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGradients.R
\name{.calculateVariogram}
\alias{.calculateVariogram}
\title{Calculate variogram-based spatial statistics}
\usage{
.calculateVariogram(object, test_mat, BPPARAM)
}
\description{
Calculate variogram-based spatial statistics
}
\keyword{internal}
