% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetabolicScores.R
\name{.filterPathwayGenes}
\alias{.filterPathwayGenes}
\title{Filter pathway genes to those present in data}
\usage{
.filterPathwayGenes(pathway_list, object, verbose = TRUE)
}
\description{
Filter pathway genes to those present in data
}
\keyword{internal}
