% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpatialMetabolic.R
\name{.filterSpatialMetabolic}
\alias{.filterSpatialMetabolic}
\title{Helper function for filtering SpatialMetabolic objects}
\usage{
.filterSpatialMetabolic(spm, min_features, min_spots)
}
\arguments{
\item{spm}{SpatialMetabolic object}

\item{min_features}{Minimum features per spot}

\item{min_spots}{Minimum spots per feature}
}
\value{
Filtered SpatialMetabolic object
}
\description{
Helper function for filtering SpatialMetabolic objects
}
\keyword{internal}
