% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetabolicScores.R
\name{.getDefaultPathways}
\alias{.getDefaultPathways}
\title{Helper to get default pathways}
\usage{
.getDefaultPathways(pathway_names, object)
}
\description{
Helper to get default pathways
}
\keyword{internal}
