% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareStates.R
\name{.hypergeometricTest}
\alias{.hypergeometricTest}
\title{Hypergeometric test for pathway enrichment}
\usage{
.hypergeometricTest(sig_genes, pathways, universe, min_genes, direction)
}
\description{
Hypergeometric test for pathway enrichment
}
\keyword{internal}
