% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareStates.R
\name{.permutationTest}
\alias{.permutationTest}
\title{Permutation test implementation}
\usage{
.permutationTest(
  test_mat,
  ref_idx,
  test_idx,
  feature_names,
  n_permutations,
  BPPARAM
)
}
\description{
Permutation test implementation
}
\keyword{internal}
