% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareStates.R
\name{.spatialLinearModel}
\alias{.spatialLinearModel}
\title{Spatial linear model (placeholder)}
\usage{
.spatialLinearModel(
  object,
  test_mat,
  conditions,
  ref_group,
  test_group,
  feature_names,
  BPPARAM
)
}
\description{
Spatial linear model (placeholder)
}
\keyword{internal}
