% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareStates.R
\name{.wilcoxonTest}
\alias{.wilcoxonTest}
\title{Wilcoxon rank-sum test implementation}
\usage{
.wilcoxonTest(test_mat, ref_idx, test_idx, feature_names, BPPARAM)
}
\description{
Wilcoxon rank-sum test implementation
}
\keyword{internal}
