% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{findMetabolicNeighborhoods}
\alias{findMetabolicNeighborhoods}
\title{Find metabolic neighborhoods}
\usage{
findMetabolicNeighborhoods(object, ...)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{...}{Additional parameters}
}
\description{
Find metabolic neighborhoods
}
