% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findVariableFeatures}
\alias{findVariableFeatures}
\title{Find variable features}
\usage{
findVariableFeatures(object, n_features = 2000, method = "vst", verbose = TRUE)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{n_features}{Number of features to select}

\item{method}{Method for finding variable features ("vst", "mvp", "disp")}

\item{verbose}{Print messages}
}
\value{
Object with variable features identified
}
\description{
Find variable features
}
