% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolicPathways.R
\name{getATPGenes}
\alias{getATPGenes}
\title{Get ATP synthesis and consumption genes}
\usage{
getATPGenes(species = "mouse")
}
\arguments{
\item{species}{Species name ("mouse" or "human")}
}
\value{
Character vector of gene symbols
}
\description{
Get ATP synthesis and consumption genes
}
