% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolicPathways.R
\name{getGlycolysisGenes}
\alias{getGlycolysisGenes}
\title{Get glycolysis pathway genes}
\usage{
getGlycolysisGenes(species = "mouse")
}
\arguments{
\item{species}{Species name ("mouse" or "human")}
}
\value{
Character vector of gene symbols
}
\description{
Get glycolysis pathway genes
}
\examples{
# Get mouse glycolysis genes
glycolysis_genes <- getGlycolysisGenes("mouse")
}
