% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolicPathways.R
\name{getMitoBiogenesisGenes}
\alias{getMitoBiogenesisGenes}
\title{Get mitochondrial biogenesis genes}
\usage{
getMitoBiogenesisGenes(species = "mouse")
}
\arguments{
\item{species}{Species name ("mouse" or "human")}
}
\value{
Character vector of gene symbols
}
\description{
Get mitochondrial biogenesis genes
}
