% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolicPathways.R
\name{getMitoDynamicsGenes}
\alias{getMitoDynamicsGenes}
\title{Get mitochondrial dynamics genes}
\usage{
getMitoDynamicsGenes(species = "mouse")
}
\arguments{
\item{species}{Species name ("mouse" or "human")}
}
\value{
Character vector of gene symbols
}
\description{
Get mitochondrial dynamics genes
}
