% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolicPathways.R
\name{getOXPHOSGenes}
\alias{getOXPHOSGenes}
\title{Metabolic pathway gene sets}
\usage{
getOXPHOSGenes(species = "mouse")
}
\arguments{
\item{species}{Species name ("mouse" or "human")}
}
\value{
Character vector of gene symbols
}
\description{
Functions to retrieve curated metabolic pathway gene sets
Get oxidative phosphorylation genes
}
\examples{
# Get mouse OXPHOS genes
oxphos_genes <- getOXPHOSGenes("mouse")
length(oxphos_genes)  # Should be ~140 genes
}
