% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolicPathways.R
\name{getSVDGenes}
\alias{getSVDGenes}
\title{Get small vessel disease signature genes}
\usage{
getSVDGenes(species = "mouse")
}
\arguments{
\item{species}{Species name ("mouse" or "human")}
}
\value{
Character vector of gene symbols
}
\description{
Get small vessel disease signature genes
}
