% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadSpatialMetabolic}
\alias{loadSpatialMetabolic}
\title{Load SpatialMetabolic object}
\usage{
loadSpatialMetabolic(file)
}
\arguments{
\item{file}{Input file path}
}
\value{
SpatialMetabolic object
}
\description{
Load SpatialMetabolic object
}
