% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{metabolicPathways<-}
\alias{metabolicPathways<-}
\title{Set metabolic pathways}
\usage{
metabolicPathways(x) <- value
}
\arguments{
\item{x}{SpatialMetabolic object}

\item{value}{List of pathways}
}
\description{
Set metabolic pathways
}
