% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{metabolicPathways}
\alias{metabolicPathways}
\alias{metabolicPathways,SpatialMetabolic-method}
\alias{metabolicPathways<-,SpatialMetabolic-method}
\title{Generic method definitions for SpatialMetabolic}
\usage{
metabolicPathways(x)

\S4method{metabolicPathways}{SpatialMetabolic}(x)

\S4method{metabolicPathways}{SpatialMetabolic}(x) <- value
}
\arguments{
\item{x}{A SpatialMetabolic object}

\item{value}{List of pathway gene sets}
}
\value{
List of pathway gene sets

Modified SpatialMetabolic object
}
\description{
This file contains all generic method definitions

Implementation of accessor methods
}
\section{Methods (by class)}{
\itemize{
\item \code{metabolicPathways(SpatialMetabolic)}: Access metabolic pathways

\item \code{metabolicPathways(SpatialMetabolic) <- value}: Set metabolic pathways

}}
