% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{metabolicScores<-}
\alias{metabolicScores<-}
\title{Set metabolic scores}
\usage{
metabolicScores(x) <- value
}
\arguments{
\item{x}{SpatialMetabolic object}

\item{value}{Matrix of scores}
}
\description{
Set metabolic scores
}
