% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{metabolicScores}
\alias{metabolicScores}
\alias{metabolicScores,SpatialMetabolic-method}
\alias{metabolicScores<-,SpatialMetabolic-method}
\title{Access metabolic scores}
\usage{
metabolicScores(x)

\S4method{metabolicScores}{SpatialMetabolic}(x)

\S4method{metabolicScores}{SpatialMetabolic}(x) <- value
}
\arguments{
\item{x}{A SpatialMetabolic object}

\item{value}{Matrix of metabolic scores}
}
\value{
Matrix of metabolic scores

Modified SpatialMetabolic object
}
\description{
Access metabolic scores
}
\section{Methods (by class)}{
\itemize{
\item \code{metabolicScores(SpatialMetabolic)}: Access metabolic scores

\item \code{metabolicScores(SpatialMetabolic) <- value}: Set metabolic scores

}}
