% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_extras.R
\name{plotMetabolicNetwork}
\alias{plotMetabolicNetwork}
\title{Plot metabolic correlation network}
\usage{
plotMetabolicNetwork(
  object,
  min_correlation = 0.3,
  layout = "fr",
  node_size = 5,
  edge_width_range = c(0.5, 3),
  colors = NULL
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{min_correlation}{Minimum correlation to show edge}

\item{layout}{Network layout algorithm}

\item{node_size}{Node size}

\item{edge_width_range}{Range for edge widths}

\item{colors}{Node colors}
}
\value{
ggplot object
}
\description{
Plot metabolic correlation network
}
