% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_extras.R
\name{plotMetabolicScatter}
\alias{plotMetabolicScatter}
\title{Plot metabolic state transitions}
\usage{
plotMetabolicScatter(
  object,
  pathway1,
  pathway2,
  group_by = "condition",
  add_ellipse = TRUE,
  point_alpha = 0.6
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{pathway1}{First pathway}

\item{pathway2}{Second pathway}

\item{group_by}{Grouping variable}

\item{add_ellipse}{Add confidence ellipses}

\item{point_alpha}{Point transparency}
}
\value{
ggplot object
}
\description{
Plot metabolic state transitions
}
