% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMetabolicSummary}
\alias{plotMetabolicSummary}
\title{Create summary plot of all metabolic pathways}
\usage{
plotMetabolicSummary(
  object,
  group_by = "condition",
  plot_type = "box",
  show_significance = TRUE
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{group_by}{Grouping variable}

\item{plot_type}{Type of summary plot ("box", "heatmap", "radar")}

\item{show_significance}{Whether to show significance annotations}
}
\value{
ggplot object
}
\description{
Create summary plot of all metabolic pathways
}
