% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_extras.R
\name{plotQCSpatial}
\alias{plotQCSpatial}
\title{Create comprehensive QC plot}
\usage{
plotQCSpatial(
  object,
  metrics = c("nCount_RNA", "nFeature_RNA", "percent_mt", "percent_ribo"),
  ncol = 2
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{metrics}{QC metrics to plot}

\item{ncol}{Number of columns}
}
\value{
ggplot object
}
\description{
Create comprehensive QC plot
}
