% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_extras.R
\name{plotSampleCorrelation}
\alias{plotSampleCorrelation}
\title{Plot sample-to-sample correlation heatmap}
\usage{
plotSampleCorrelation(
  object,
  use_pathways = TRUE,
  method = "pearson",
  cluster = TRUE
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{use_pathways}{Whether to use pathway scores or genes}

\item{method}{Correlation method}

\item{cluster}{Whether to cluster samples}
}
\value{
pheatmap or ggplot object
}
\description{
Plot sample-to-sample correlation heatmap
}
