% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSpatialGradients}
\alias{plotSpatialGradients}
\title{Plot spatial gradients}
\usage{
plotSpatialGradients(
  object,
  top_n = 10,
  result_type = "scores",
  show_names = TRUE,
  name_size = 3
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{top_n}{Number of top features to plot}

\item{result_type}{Which gradient results to use ("scores" or "genes")}

\item{show_names}{Whether to show feature names}

\item{name_size}{Size of feature names}
}
\value{
ggplot object
}
\description{
Plot spatial gradients
}
