% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_extras.R
\name{plotSpatialGrid}
\alias{plotSpatialGrid}
\title{Additional visualization functions for SpatialMetabolics}
\usage{
plotSpatialGrid(
  object,
  features,
  sample = NULL,
  ncol = 4,
  point_size = 0.8,
  sync_scale = TRUE,
  color_palette = "viridis"
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{features}{Features to plot in grid}

\item{sample}{Sample to plot}

\item{ncol}{Number of columns}

\item{point_size}{Point size}

\item{sync_scale}{Synchronize color scales}

\item{color_palette}{Color palette}
}
\value{
ggplot object
}
\description{
Additional visualization functions for SpatialMetabolics
}
