% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessSpatial}
\alias{preprocessSpatial}
\title{Preprocess SpatialMetabolic object}
\usage{
preprocessSpatial(
  object,
  normalize_method = "logNormalize",
  n_variable_features = 2000,
  n_pcs = 30,
  verbose = TRUE
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{normalize_method}{Normalization method}

\item{n_variable_features}{Number of variable features}

\item{n_pcs}{Number of principal components}

\item{verbose}{Print messages}
}
\value{
Preprocessed object
}
\description{
Preprocess SpatialMetabolic object
}
