% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runUMAP}
\alias{runUMAP}
\title{Run UMAP on spatial data}
\usage{
runUMAP(
  object,
  n_components = 2,
  use_dims = 30,
  n_neighbors = 30,
  min_dist = 0.3,
  metric = "euclidean",
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{n_components}{Number of UMAP components}

\item{use_dims}{Number of PCs to use}

\item{n_neighbors}{Number of neighbors}

\item{min_dist}{Minimum distance}

\item{metric}{Distance metric}

\item{seed}{Random seed}

\item{verbose}{Print messages}
}
\value{
Object with UMAP results
}
\description{
Run UMAP on spatial data
}
