% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveSpatialMetabolic}
\alias{saveSpatialMetabolic}
\title{Save SpatialMetabolic object}
\usage{
saveSpatialMetabolic(object, file, compress = TRUE)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{file}{Output file path}

\item{compress}{Whether to compress the file}
}
\description{
Save SpatialMetabolic object
}
