% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,SpatialMetabolic-method}
\alias{show,SpatialMetabolic-method}
\title{Show method for SpatialMetabolic objects}
\usage{
\S4method{show}{SpatialMetabolic}(object)
}
\arguments{
\item{object}{A SpatialMetabolic object}
}
\description{
Show method for SpatialMetabolic objects
}
