% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{spatialGraphs<-}
\alias{spatialGraphs<-}
\title{Set spatial graphs}
\usage{
spatialGraphs(x) <- value
}
\arguments{
\item{x}{SpatialMetabolic object}

\item{value}{List of graphs}
}
\description{
Set spatial graphs
}
