% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{spatialGraphs}
\alias{spatialGraphs}
\alias{spatialGraphs,SpatialMetabolic-method}
\alias{spatialGraphs<-,SpatialMetabolic-method}
\title{Access spatial graphs}
\usage{
spatialGraphs(x)

\S4method{spatialGraphs}{SpatialMetabolic}(x)

\S4method{spatialGraphs}{SpatialMetabolic}(x) <- value
}
\arguments{
\item{x}{A SpatialMetabolic object}

\item{value}{List of spatial graphs}
}
\value{
List of spatial graphs

Modified SpatialMetabolic object
}
\description{
Access spatial graphs
}
\section{Methods (by class)}{
\itemize{
\item \code{spatialGraphs(SpatialMetabolic)}: Access spatial graphs

\item \code{spatialGraphs(SpatialMetabolic) <- value}: Set spatial graphs

}}
