% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SpatialMetabolic-class}
\alias{SpatialMetabolic-class}
\title{SpatialMetabolic Class}
\description{
S4 class for spatial metabolic analysis extending SpatialExperiment
}
\section{Slots}{

\describe{
\item{\code{metabolic_pathways}}{list of metabolic pathway gene sets}

\item{\code{spatial_graphs}}{list of spatial neighborhood graphs}

\item{\code{metabolic_scores}}{matrix of pathway activity scores}

\item{\code{analysis_results}}{list of analysis results}

\item{\code{comparison_groups}}{factor defining comparison groups}
}}

