% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetSpatial}
\alias{subsetSpatial}
\title{Subset SpatialMetabolic object}
\usage{
subsetSpatial(object, cells = NULL, features = NULL, samples = NULL)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{cells}{Cells to keep}

\item{features}{Features to keep}

\item{samples}{Samples to keep}
}
\value{
Subsetted object
}
\description{
Subset SpatialMetabolic object
}
