% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toSeurat}
\alias{toSeurat}
\title{Export SpatialMetabolic to Seurat}
\usage{
toSeurat(object, assay_name = "RNA", add_metabolic_scores = TRUE)
}
\arguments{
\item{object}{SpatialMetabolic object}

\item{assay_name}{Name for the RNA assay}

\item{add_metabolic_scores}{Whether to add metabolic scores as metadata}
}
\value{
Seurat object
}
\description{
Export SpatialMetabolic to Seurat
}
